First commit
I have had a confusion on control theory, specifically on the second-order delay element. I thought that if the bode diagram had no peak gain, the step response exhibit no overshoot. But, the truth is not like this. With the damping factor zeta between 0.707 and 1, the bode diagram has no peak gain wheareas the step response exhibits overshoot.